Add the three pieces of legislation that go with the music
Add the basic legislation that goes with the music: The Share, Voting Rights for All, and the Fair Commerce Act.

Ended up also putting up a fair amount of Attorney General Reform, including the constitutional amendment and a landing/discussion page.

Reoorganized as lot of the directories to be more consistent.

Ran internal link checking across the whole site.

Fixed the website to scroll both the table of contents and the main content and to adapt that to a single column on phones.