Change the wording on Duty of No Harm to Duty of Fairness
The existing framework in contract law is not "No Harm," which is potentially too full of liability.

The existing framework, around for centuries, is  "Good Faith and Fair Dealing"

So rather than get stuck in contentious debate about foreseeing harm, lets just go to the existing language for contracts.

And this should work as well. The fundamental argument we are making is that contracts are being used by fiduciaries to shield beneficiaries from the requirement of fair dealing of stakeholders, but creating a middle-man.

And change the example on the landing page to clarify it's the beneficiary's company getting the shield through a contract, and change the industry of the beneficiary from waste disposal themselves, with a sub-contract for specific waste, to a manufacturer who has a contract for toxic waste (because this should be more clear).